First pass of networking support
Basic URL fetching works on Linux and PS3. Compiles on Windows
but is not working until we can initialize Winsock.

Move object watcher from lbshell into base.

Modify gyp files to reference googleurl/googleurl.gyp rather than
build/temp_gyp/googleurl.gyp, and third_party/openssl.

Disable net_resources actions for Cobalt- we don't have tools/grit,
and I don't think net_resources is needed.

Set use_native_http_stack variable default to 0.

Some DIAL compilation fixes for Linux- don't use file-scope
std::strings as they require global destructors to run.

Fix a few compiler warnings in under MSVC related to
signed/unsigned conversion and type narrowing.

Add actual_target_arch to filename_rules.gypi, to allow
building lbshell/src/*_win.cc on Windows.

Fix a few lint errors.

+++ net/http/http_version.h
@@ -16,7 +16,8 @@ class HttpVersion {
   HttpVersion() : value_(0) { }
 
   // Build from unsigned major/minor pair.
-  HttpVersion(uint16 major, uint16 minor) : value_(major << 16 | minor) { }
+  HttpVersion(uint16 major, uint16 minor)
+      : value_(static_cast<uint32>(major) << 16 | minor) { }
 
   // Major version number.
   uint16 major_value() const {
